hard: preparation for CI/CD and fix bug orders/summary api

+++ order-service/src/main/java/com/moreira/order_service/models/PriceSummaryRecord.java
@@ -2,5 +2,5 @@ package com.moreira.order_service.models;
 
 import java.math.BigDecimal;
 
-public record PriceSummaryRecord (String customer, Double total){
+public record PriceSummaryRecord (String customer, BigDecimal total){
 }
